+++ lucene/core/src/java/org/apache/lucene/document/LatLonPointDistanceQuery.java
@@ -247,7 +247,7 @@ final class LatLonPointDistanceQuery extends Query {
           DocIdSetBuilder.BulkAdder adder;
 
           @Override
-          public void grow(int count) {
+          public void grow(long count) {
             adder = result.grow(count);
           }
