+++ lucene/core/src/java/org/apache/lucene/document/SpatialQuery.java
@@ -413,7 +413,7 @@ abstract class SpatialQuery extends Query {
       DocIdSetBuilder.BulkAdder adder;
 
       @Override
-      public void grow(int count) {
+      public void grow(long count) {
         adder = result.grow(count);
       }
