+++ lucene/core/src/java/org/apache/lucene/search/comparators/NumericComparator.java
@@ -231,7 +231,7 @@ public abstract class NumericComparator<T extends Number> extends FieldComparato
             DocIdSetBuilder.BulkAdder adder;
 
             @Override
-            public void grow(int count) {
+            public void grow(long count) {
               adder = result.grow(count);
             }
